+++ docmind-api-20220711/ChangeLog.md
@@ -1,3 +1,8 @@
+2025-10-27 Version: 1.4.12
+- Update API GetDocParserResult: add request parameters ExcludeFields.
+- Update API SubmitDocParserJob: add request parameters NeedHeaderFooter.
+
+
 2025-09-12 Version: 1.4.11
 - Update API SubmitDocParserJob: add request parameters CustomOssConfig.
 - Update API SubmitDocParserJob: add request parameters LLMParam.